server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Tae Hyoung Jo <taehyoung.jo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.110.0"}
psycopg = "^3.1.18"
pyvespa = "^0.39.0"
ipykernel = "^6.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
